Add bookings

+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -1,23 +1,23 @@
 package ru.practicum.shareit.item.dto;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ItemDto {
-    private int id;
+    private Long id;
     @NotBlank(message = "Имя не может быть пустым")
     private String name;
     @NotBlank(message = "Описание должно быть заполнено")
     private String description;
     @NotNull(message = "Необходимо указать доступность вещи")
     private Boolean available;
-    private int ownerId;
-    private int request;
+    private Long ownerId;
+    private Long request;
 }
