+++ src/main/java/ru/practicum/shareit/user/UserDto/UserDto.java
@@ -1,22 +1,22 @@
 package ru.practicum.shareit.user.UserDto;
 
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
 import java.util.Objects;
 
 @Data
 @NoArgsConstructor
 public class UserDto {
-    private Integer id;
+    private Long id;
     private String name;
     @NotBlank(message = "Поле email не должно быть пустым")
     @Email(message = "Поле email должно быть заполнено как адрес электронной почты")
     private String email;
 
-    public UserDto(Integer id, String name, String email) {
+    public UserDto(Long id, String name, String email) {
         this.id = id;
         if (name.isBlank()) {
             this.name = email;
